Add RNG distributions and voxel type indexing
This commit adds a richer random-number API for continuous and discrete distributions, including normal, gamma, beta, Poisson, Zipf, and weighted sampling. It also updates voxel type IDs to usize and introduces a voxel type index registry for assigning and fetching types by ID. Generated target artifacts were cleaned up, and a .gitignore was added.

+++ src/World/voxel_type.rs
@@ -1,7 +1,7 @@
 use crate::World::voxel_components::VoxelComponent;
 
 struct VoxelType {
-    id: u32,
+    id: usize,
     name: String,
     properties: Vec<VoxelComponent>,
 }
